Python/urlshortener/requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.0
colorama==0.3.9
decorator==4.0.11
Django==1.11.2
django-extensions==2.0.6
django-rest-framework-mongoengine==3.3.1
django-test-addons==1.0
djangorestframework==3.6.3
docker==2.3.0
docker-compose==1.14.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
factory-boy==2.8.1
Faker==0.7.17
functools32==3.2.3.post2
ipaddress==1.0.18
ipython==5.4.1
ipython-genutils==0.2.0
jsonschema==2.6.0
mongoengine==0.13.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pymongo==3.4.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.11.1
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
texttable==0.8.8
traitlets==4.3.2
typing==3.6.4
wcwidth==0.1.7
websocket-client==0.44.0
